add try_from for ParquetMetaData
